+++ sdks/go/node/containerruntime/docker/runContainer_test.go
@@ -241,7 +241,7 @@ var _ = Context("RunContainer", func() {
 
 			expectedNetworkingConfig := &network.NetworkingConfig{
 				EndpointsConfig: map[string]*network.EndpointSettings{
-					dockerNetworkName: {
+					networkName: {
 						Aliases: []string{
 							*providedReq.Name,
 						},